# project-pipeline/backend/openapi.yaml
openapi: 3.0.0
info:
  title: Project Pipeline Management API
  description: |
    AI-powered Erasmus+ project planning and management system.

    Features:
    - 6 specialized AI agents for different planning domains
    - Real-time budget tracking and alerts
    - Automated vendor communications
    - Learning system that improves with each project
    - Erasmus+-compliant reporting

  version: 1.0.0
  contact:
    name: API Support
    email: support@projectpipeline.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.projectpipeline.com
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Vendor:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        website:
          type: string
        location:
          type: string
        rating:
          type: number
          format: float
paths:
...
  /vendors:
    get:
      tags: [Vendors]
      summary: List vendors
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of vendors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Vendor'
    post:
      tags: [Vendors]
      summary: Create vendor
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, type, email]
      responses:
        '201':
          description: Vendor created

  # Communications
  /communications/quote-request:
    post:
      tags: [Communications]
      summary: Send quote request to vendor
      description: AI composes and sends email
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vendorId, phaseId]
              properties:
                vendorId:
                  type: string
                phaseId:
                  type: string
                customMessage:
                  type: string
      responses:
        '200':
          description: Quote request sent
  /communications/bulk-quote-request:
    post:
      tags: [Communications]
      summary: Send quote requests to multiple vendors
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vendorIds, phaseId]
              properties:
                vendorIds:
                  type: array
                  items:
                    type: string
                phaseId:
                  type: string
      responses:
        '200':
          description: Bulk requests sent